requirements.txt
FastAPI
python-jose
requests
auth0-python
uvicorn
pydantic
python-dotenv